pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "streamsentinel"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    # Kafka
    "confluent-kafka==2.3.0",
    # Stream processing
    "faust-streaming==0.10.14",
    # ML
    "torch==2.2.0",
    "numpy==1.26.4",
    "scikit-learn==1.4.0",
    # API
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "websockets==12.0",
    # Storage
    "redis==5.0.1",
    "asyncpg==0.29.0",
    "sqlalchemy[asyncio]==2.0.28",
    "alembic==1.13.1",
    # Validation
    "pydantic==2.6.3",
    "pydantic-settings==2.2.1",
    # Observability
    "prometheus-client==0.20.0",
    # Utilities
    "faker==23.3.0",
    "python-dotenv==1.0.1",
    "mlflow==2.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.2",
    "pytest-asyncio==0.23.5",
    "httpx==0.27.0",
    "ruff==0.3.0",
]